Create lint_python.yml
As requested at https://github.com/robotframework/robotframework/pull/3313#issuecomment-606458304  This is a exact replica of #3421
