+++ atest/testdata/libdoc/Annotations.py
@@ -1,3 +1,6 @@
+# flake8: noqa
+
+
 class UnknownType(object):
     pass
