+++ src/robot/htmldata/jsonwriter.py
@@ -13,7 +13,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from robot.utils import PY2
+from robot.utils.robottypes import long, unicode
 
 
 class JsonWriter(object):
@@ -75,7 +75,7 @@ class _Dumper(object):
 
 
 class StringDumper(_Dumper):
-    _handled_types = (str, unicode) if PY2 else str
+    _handled_types = (str, unicode)
     _search_and_replace = [('\\', '\\\\'), ('"', '\\"'), ('\t', '\\t'),
                            ('\n', '\\n'), ('\r', '\\r'), ('</', '\\x3c/')]
 
@@ -91,7 +91,7 @@ class StringDumper(_Dumper):
 
 class IntegerDumper(_Dumper):
     # Handles also bool
-    _handled_types = (int, long) if PY2 else int
+    _handled_types = (int, long)
 
     def dump(self, data, mapping):
         self._write(str(data).lower())
